Implement a basic working example of Record types.
The implementation leaves a lot to be desired, but it's roughly
the shape I had in mind. It could probably benefit from a few extra
safety checks...

This commit also includes a pretty major refactor of the ComponentInterface
type and its members, to make it easier to generate the bindings.