Expose bindings generation in a nicer way.
In addition to producing a compiled library with the component code,
each uniffi component now also produces a helper executable. This
executable embeds the IDL definition as it was at the time of generation
and can be used to generate compatible bindings for target foreign
languages.

The helper executable it also a convenient way to execute scripts in
the foreign language with the component configured and available.
We can use this as the beginnings of a simple test harness.