Cleanup and document the process for generating the foreign language bindings.
I've pinched a little trick from wasm-bindgen that I think is quite neat and
that gives a nice developer experience - when compiling the rust code down to
a shared library, we now embed a serialization of the `ComponentInterface` in a
custom section of the object file. This makes the file self-describing, since
an external tool can parse the `ComponentInterface` out of the library itself
in order to generate bindings that work with it.

Using this technique, there is now a top-level `uniffi-bindgen` command that
will generate bindings from any given uniffi-compiled library file. I've also
kept the existing developer experience of being able to `cargo run generate`
within the component crate since it simplifies the most common workflow.

In the future, you could imagine that e.g. the python binding would not need
to ahead-of-time generate any code at all, but could instead load the compiled
library, inspect and parse the `ComponentInterface` section, and dynamically
generate the API described therein. But that's for a far, far future!

(Why yes, this is very similar to how the Wasm Interface Types proposal works
on top of core Wasm, thanks for noticing...)