Turn example scripts into simple test fixtures.
Prior to this commit, the only "tests" we had for the generated code
were a small set of example scripts in the various target languages,
and we could "test" them by manually executing the script and seeing
whether it printed plausible output.

With this commit, those scripts can now be proper testcases. It has
involved a fair bit of restructuring, but broadly:

* The crate and its examples are now part of a single workspace,
  so `cargo test` in the top-level directory will test them all.
* There is a new `uniffi_macros` crate with a single macro called
  `build_foreign_language_testcases!`. Given a list of script files,
  it will generate a testcase shim for each one, executing the file
  via an appropriate interpret and failing the test if it does not
  exit successfully.
* The various example crates have had their example scripts converted
  into testcases in this manner.