Clarify type conversions by differentiating between API-level and FFI-level types.
Previously, we had a single set of types (the `TypeReference` enum) which were
used both in API function signatures, and in the signatures of internal FFI functions.
This was nice and simple to get started but involved a lot of duplication of the
implicit conversions that happen when sending data across the FFI.

This commit splits the types into two different...um...types. We have the `Type`
enum to represent the rich high-level types that appear in the component API, and
a separate `FFIType` enum to represent the much more constrained set of types that
are used to pass data back-and-forth across the FFI.

To prove out the approach, this commit also removes the asymmetry in how strings
are handled in argument position vs return position. The API-level `Type::String`
always corresponds to an owned Rust string, and is passed across the FFI as a pointer
represented by the low-level `FFIType::RustString`. Foreign language code can create
a `FFIType::RustString` using a new helper function that takes `FFIType::ForeignStringRef`
and copies it into the rust world.

The result is a smaller number of moving part in the foreign language binding
generation.