Use our own bindgen when building our crates, rather than system bindgen.
When working on uniffi itself, it's very inconvenient to have to rely
on an externally-installed `uniffi-bindgen` command in order to run the
tests for the in-tree example crates. You basically need to remember to
manually `cargo build -p uniffi_bindgen` anytime you make a change to the
bindings genertor, or the tests won't test what you think they should.

This commit adds a feature-flag to the `uniffi` and `uniffi_build` crates
so that they will use the builtin `unifi_bindgen` module from this workspace
rather than whatever one they find on the system. I think this gives a much
better development experience when working on this repo itself, in particular
by allow `cargo test` to Just Work (tm) and always reflect your latest changes.

On the other hand, it means the examples in this repo are doing a thing that
we might not expect actual consuming crates to do. I think it's a good tradeoff
on balance though.