Clarify terminology around passing values across the FFI.
Previously, we used a loosely-defined mix of terms like "lower" and "lower into"
to describe passing data across the FFI in two very different ways (the former
by converting to a primitive C type, the later by serializing to a bytebuffer).
Sometimes when we said "lower" we meant the former, and sometimes we meant the
later, and this was not consistent between the different codegen backends.

This commit tries to clean up the terminology and bring the Rust, Kotlin and
Swift code generators into agreement on which activities are called what.
Broadly:

* "Lowering" is now exclusively about converting from a high-level language
  type into a low-level C-compatible type, and "lifting" is the reverse.
* Serializing into a bytebuffer is now called "writing" and deserializing
  from a bytebuffer is now called "reading".

I've also consolidated all these into a single trait/protocol/what-have-you
where possible, since we've had confusion in the past where some types implemented
e.g. lowering but not writing.

I haven't touched the python bindings, because I'm working on a bigger
refactor in a separte commit.