Clean up the README, move some scene-setting stuff into ADRs.
The README currently contains a lot of scene-setting motivational and
aspirational material, such as a lot of words about how great I think
WASM Interface Types are and how we should deliberately model our work
after them. This was written for the purposes of explaining the "what"
and "why" of the initial prototype, and it was aimed internally, to
convince folks that we should build this thing. Now that we are actually
building this thing, the README should focus more on people who might
want to *use* this thing, and in its current form I think it buries the
key details in a wall of implementation notes.

This commit revamps and significantly shortens the README, to make
it more relevant to potential consumers of this tool.

However, that doesn't mean we should throw that motivational material
away! Instead, let's move it into an explicit historical record of
what we're doing here and why. Several teams at Mozilla have been using
"Architectural Decision Records" for this purpose, so I figure it'll
be worth us trying that format here.