Remove `cargo_metadata` and make other versions match m-c's.
When building in Gecko, `cargo_metadata` panics in a macro. This is
fixed in oli-obk/cargo_metadata#123 (and in Serde upstream), but
pulling `uniffi` in as a path dependency in Gecko, and depending on
`cargo_metadata` as a Git dependency, confuses Cargo. It doesn't
update the vendored versions of Serde or other dependencies, then
fails to build complaining it can't satisfy the newer versions that
`uniffi` requests.

In the interest of getting a demo working, I've just commented out
the dependencies for now, and downgraded the other versions to match
what's already vendored in m-c.