Pass strings as `RustBuffer` structs, rather than `char*` arrays.
Null-terminated strings are nice when writing APIs by hand, because
they're convenient. But they come with several downsides:

* Overhead of scanning for the terminating null character.
* Incapable of representing strings containing null bytes.

We're not writing APIs by hand, so we don't really get the benefit
of convenience and are left only with the downsides.

This commit switches `<String as uniffi::ViaFfi`> to lower into
a `RustBuffer` by directly exposing the underlying `Vec<u8>` from
the string, rather than its previous behaviour of converting into
a `CString`. It also removes some of the string-handling helpers
from the FFI layer since they're no longer necessary.

We do need to keep one string-handling helper, at least for now.
The `ExternError` struct uses a `const char*` for its `message`
proprty, and the foreign-language code needs to be able to call
a destructor to free it. We could make our own `ExternError` that
passes the message in a `RustBuffer` to avoid this, but that's
left to followup work.