Include a short checksum in generated FFI function names, for extra link-time safety.
Since the foreign-language bindings are generated separately from the compiled Rust scaffolding,
user error might lead to a situation where the bindings try to link against an incompatible
version of the compiled code. With luck this would produce a link-time error, but it might
also lead to arbitrary memory unsafety, depending on what's changed between the two sides
of the FFI boundary.

This commit attempts to add an additional layer of safety to prevent the foreign-language
bindings from loading incompatible Rust code. We insert a short checksum into the name of
every single FFI function, which will change if the WebIDL interface changes or if the version
of uniffi changes. Theory being, the bindings can't load and call incompatible FFI functions
if they don't know what they're called!

With this in place, attempts to use foreign-language bindings generated from a different
IDL than the Rust code, or with a different version of uniffi, should fail with an inscrutable
link-time error. But at least they'll fail early and safely.

This may also cause some failures in cases that would have been safe, such as if the two
sounds of the boundary were generated using two different versions of Rust. But if that does
somehow happen, it probably indicates a fragile build process that's at risk of bigger
incompatible probablems, so catching it early seems worthwhile.