Pass a `Context` struct to Gecko JS filters.
There are two pieces of information that we want to pass along to filters:

* The `definition_prefix` config option, used to prefix the generated
  namespace, interface, enum, and dictionary names.
* The `ffi_namespace`, used to generate the names for the `RustBuffer`,
  `RustError`, and `ForeignBytes` structs.

Threading these arguments through Askama is tricky. For example, if the
`type_ffi` filter needs to know the `ffi_namespace`, that infects any
filters that call it (like `lift_cpp` and `lower_cpp`), all the way up
the call chain. The same happens for the `definition_prefix`, which
changes the names of the types declared in the WebIDL...so now all type
declaration filters need to know about the prefix.

Making this work in the generated code is difficult, so we take the
least-worst possible option of passing around a `Context` argument
(other, less-generic names welcome!) with all the names we need.