Special-case "optionals with a default value" in Gecko.
These types are different than optionals without a default value.
Their C++ type is just `T`, not `Optional<T>`, since the default value
is a fallback if the caller doesn't pass one. But they must be
serialized as if they were `Optional<T>`, because the Rust side of the
FFI will try to deserialize them as optionals instead of instances of
the value.