Flush struct memory when writing to a RustBuffer from Kotlin.
JNA structs have both JVM-level fields and an allocated slab of native
memory, and values need to be kept in sync between the two. This is
handled automatically by JNA when passing structs to native functions,
but our RustBufferBuilder might modify and return a struct without
passing it to native code.

This commit adds some explicit use of `writeField` to ensure that
the values from JVM-level fields are synchronized to native memory.