Don't use i32 helpers in generated Enum code.
Previously, the generated Kotlin code for enums would try to use
helpers from the `Int` companion class, but we would only emit the
code for those helpers if the interface was also using the integer
type.

Now, the generated Kotlin code for enums directly uses lower-level
methods for reading and writing integers, so that it doesn't depend
on other generated code.

Fixes #356.