Add callback interfaces for Kotlin (#344) r=rfk
* Add callback interfaces for Kotlin

We'd like to have a Rust trait:

```rust
trait FetchClient {
    fn fetch(url: String) -> String
}

// later
impl Nimbus {
    fn set_fetch_client(client: Box<dyn FetchClient>) {}
}
```

Declared in the IDL:

```idl
interface Nimbus {
    set_fetch_client(FetchClient client);
    update_experiments();
}

callback interface FetchClient {
    string fetch(string url);
};
```

And used in Kotlin:

```kotlin
class ConceptFetchClient: FetchClient {
    override fun fetch(url: String): String {
        ...
    }
}

nimbus.setFetchClient(ConceptFetchClient())
nimbus.updateExperiments() // uses ConceptFetchClient

```

This commit sends JNA Callbacks to send a callback interface to Rust (a.k.a. `ForeignCallback`).

This implementation of the `ForeignCallback` is individual to the callback interface— on the kotlin side; in this case this is called: `CallbackInterfaceFetchClientFFI`.

As the `ConceptFetchClient` is sent from kotlin, it is placed in a kotlin `ConcurrentHandleMap`.

A Rust `impl` of corresponding `FetchClient` trait is generated which proxies all calls through the `ForeignCallback` to kotlin, i.e. `CallbackInterfaceFetchClientProxy` in Rust calls to `CallbackInterfaceFetchClientFFI` in kotlin.

The callback object i.e. the `ConceptFetchClient` is then found, and then methods are called from `CallbackInterfaceFetchClientFFI` to the client code `ConceptFetchClient`.

* Address first pass review comments

* Address reviewer comments

* Review 3 comments

* Final review comments