Update to the latest release of weedle.
The new version of weedle tracks an update in the WebIDL spec [1]
that renamed `void` to `undefined`, which makes sense for the web but
is a bit of a pain for us. To avoid churning consumers of uniffi,
this commit adds some special-case handling for the identifier `void`
in return position. You can tell that the special-casing works
by the fact that this doesn't require changes to any of our tests,
many of which use the `void` syntax.

[1] Ref https://github.com/rustwasm/weedle/pull/43