Hide Rust scaffolding from `cargo doc`.
I've been experimenting with using `cargo doc` as the canonical source
of documentation for a UniFFI component, and I quite like it so far,
except for the way that it currently documents all the autogenerated
helper functions and traint impls. So, let's hide those using the
nice `#[doc(hidden)]` annotation.