Add config option to set the cdylib name for Kotlin.
When building uniffi components in the application-services monorepo,
we know that they're all going to want to load their FFI symbols from
the combined "libmegazord.so". This adds a config option so that
the generated Kotlin code can use this cdylib name by default, rather
than having to specify it as an override at runtime.