New 'coverall' fixture to test coverage and edge-cases.
This is similar to our examples, but it's intended to be contrived and to
ensure we get good test coverage of all possible options - particularly for
future changes.

It's in the "fixtures" directory so it doesn't even need to make a cursory
effort to be a "good example"; it intentionally panics, asserts params are
certain values, has no-op methods etc. If you're trying to get your head
around uniffi then the "examples" directory will be a much better bet.