Allow objects to be returned - they can already existing in dicts.
This is just the rust side of the world, and functions must
return `std::sync::Arc<Self>` rather than plain `Self` - not
sure if that's a bug (not seamless and natural) or a "feature"
(by closing existing `Arc`s offers a path to an "identity" concept)

Note that it would be fairly simple to allow objects returned
by Rust to not be `Arc<>` because `Arc<T>` automatically supplies
`From<T>` - so we just need to generate a few `.into()` calls.
Obviously though, this will create a new `Arc` and not a clone of
an existing one, so identity will be lost.

It also more difficult to fix the other direction (ie, when the
FFI layer has a ptr to an `Arc<T>` but the rust code expects just `&T`)

It's possibly doable, but as above, it's not clear whether we actually
*want to* do that - asking the author to explicitly manage `Arc<>`s
seems less likely to cause surprises down the track - we made that
mistake before with Mutexes.