Add a test fixture for Kotlin "experimental unsigned types" warnings.
Previously, our test running infrastructure would specify the Kotlin option
 "-Xopt-in=kotlin.ExperimentalUnsignedTypes" when compiling the uniffi bindings,
in order to silence warnings about using the experimental unsigned types.

This is *not* how we intend to compile the generated bindings in practice.
Rather, we want to only opt in to experimental types at the level of the final
consumer, with all intermediate layers propagating the experimental types
annotation to the layer above them.

This commit removes the unsigned types opt-in when compiling bindings for
tests, and replaces it with `-Werror` so that any warnings when compiling
the bindings will fail the corresponding test.  (It keeps the opt-in when
running the actual test script, since that is the uppermost layer and is
where we intend real consumers to specify the opt-in behaviour).

It also adds a new test fixture that deliberately uses a bunch of unsigned
types, and tests that the Kotlin bindings compile and run successfully.

This test currently fails, because we're not propagating the annotations
correctly. So this commit should not be merged as-is.

Making the test pass is issue #444.