Implement `[ByArc]` annotation for methods on threadsafe objects.
As of ADR-0005, the use of `Arc<T>` for passing object instances will
be part of the official Rust-facing  API of UniFFI rather than being
a private implementation detail. A full implementation of that ADR is
still in progress, but in the meantime we can get a relatively easy
implementation of one small part of it: support for `Arc<Self>` as
method receiver.

This commit implements support for the `[ByArc]` annotation on method
definitions, indicating they the underlying Rust code will accept
`self: Arc<Self>` rather than `self: &Self` as method receiver. This
only works for `[Threadsafe]` interfaces since they are the only ones
that use `Arc<Self>` internally, but that's OK because we intend to
require that all interfaces are threadsafe in the near future.

The implementation is small, but not particularly pretty. I think
it's fine to push forward with given our broader plans to simplify
away a lot of the branching in this code.