Fix data race between Kotlin method calls and destruction.
In the previous implementation, a concurrent method call and call
to `destroy()` could result in the Rust code trying to access a
freed handle in the handlemap. This is caught cleanly by the handlemap,
but only as a last resort; it should never be allowed to happen in
practice.

This commit changes the way destruction works on the Kotlin objects
to use a reference count, avoiding freeing the underlying Rust object
until all method calls using it have completed.

Fixes #457.