Start keeping a changelog.
If we're going to be issuing deprecation notices and such, we really ought to
be keeping a changelog. This commit adds one, with some retroactive notes from
the `v0.9.0` release.

I had hoped to be able to automatically manage updating the section headers
in the changelog using `cargo release`, but that doesn't seem to work as I
hoped when releasing multiple crates in a single workspace. Still, let's
organize it as though we'll be using `cargo release` for this one day.