Restructure Swift bindings for easier megazording.
This re-arranges the way the Swift bindings are generated
to make a bit more of a deliberate split between the low-level
and high-level code. The intention is to make it easier for
us to compile the Rust code for multiple components together
into a single build artifact, while distributing the Swift
code as individual Swift package artifacts.

I've also taken the opportunity to add some more general docs
on how the Swift backend works.

Fixes https://mozilla-hub.atlassian.net/browse/SYNC-2679.