Refactor ViaFfi to use an associated type
No functional changes yet -- this change is mostly just replacing `Self`
with `Self::RustType`.  The only actual code change was to shift the
responsibility of lowering rust errors from the `rustcalls` module to
the code that calls into it.

This opens the door for handling external types because we can define
our own struct that implements ViaFfi for the external type.