New subcommand `generate-api` to generate all the expected Rust boilerplate from UDL
The conversion from the UDL definition to the actual Rust signature is
not always obvious, especially if just getting started.
Optional values, throwable exceptions, `Arc` vs. `&self`, ...

Given we already have the UDL and we can parse that and we do know what
code we expect ... why don't we "just" generate it?

That's what this code adds: a new Rust backend for the code generator.
It generates all structs, (error) enums, functions, and methods
(including the constructor ones) in Rust code.
All implementation is stubbed out with `todo!()`.
But because this is not really a backend it gets its own subcommand.
Its code however lives on as a `binding` because that's where it neatly
fits in.

Someone wanting to use UniFFI can start with definining the API in the
UDL file.
Then run `uniffi-bindgen generate-api path/to/src/namespace.udl`,
take the code from the generated `namespace_uniffi.rs` and start
implementing functionality.

Note: this is not intended as a full-blown backend.
It's not capable of updating user code. Once generated the expectation
is that the user copies over the function signatures to their own code.