Add a re-usable abstraction for iterating over contained types.
We currently have several methods that want to be able to recurse
through all the types contained within a particular item, such as
`type_contains_object_references` and `type_contains_unsigned_types`.
These methods duplicate the logic for recursing into our various
types of struct and are a bit weird to use.

This commit proposes a cleanup to make these methods appear more
consistent to the caller, and reduce code duplication in their
implementation.

The core idea is a new trait `IterTypes`. It can be implemented by
anything that contains instances of the `Type` enum, and provides a
single method `iter_types` to iterate over the the contained types.
It's designed to encapsulate the recursion logic, e.g. the `IterTypes`
impl for `Object` knows that it needs to recurse into each of the
methods, the `IterTypes` impl for a `Method` knows that it needs
to recurse into each argument and return type, etc.

On top of this, the `CompontentInterface` can provide a generic
implementation of methods like `item_contains_object_references` and
`item_contains_unsigned_types` that will work for any `IterTypes`. I
think this provides a nice little increase in consistency with how
those methods are implemented and called.