Drop ffi-support dependency
This seems like a nice little cleanup.  The `rustcalls` mod was importing
`IntoFfi`, but only using a trivial part of the interface.  Now we just
define our own trait to do that.

I was a slightly worried about this when I initially implemented the
`rustcalls` module, because I was concenced about uniffi consumers
defining their own `ViaFfi` implementation with some other `FfiType`.
However I now realize this is impossible, we have a `FfiType` enum in
the scaffolding code that spells out the possible types.

The last bit of functionality was the panic handling hook, I just copied
the code for that.