Use fully-qualified type names in generated scaffolding code.
This fixes #1015, which was a regression in the v0.13.0 release.

Where possible we should be robust against changes in the details of
the Rust code that implements a UniFFI component, but the generated
Rust scaffolding accidentally started assuming that the `HashMap` type
would be in scope for any interface using a map. This turned out not
to be true in practice.

To improve things, this commit makes the generated scaffolding use
fully-qualified type names for `HashMap`, `Vec` and `Option`. There
is probably more we could do on this front (e.g. the `String` type)
but these were low-hanging fruit that fixed a concrete regression,
so that's where I started.