Refactoring lift/lower/read/write/coerce for the Python bindings
Added FfiConverter classes for each type we handled.  These replace
various template filter functions:
  - coerce_py/lower_py -> [FfiConverterForType].lower()
  - lift_py -> [FfiConverterForType].lift()
  - read_py -> [FfiConverterForType].read()
  - write_py -> [FfiConverterForType].write()

I think this system is a easier to understand because most of the
type-specific code is now in one place.  It could also simplifies the
template generation.

Standardizing FfiConverter also helps out with wrapping and external
types.  For example we can now handle a function that directly returns
an external type.