Enhance external/wrapper types example and fixtures
I took this changeset from mhammond's branch to see if we could handle
it.  Most things worked, the one issue was `RustBuffer`.  With external
types we have a weird situation where both libraries define
`RustBuffer`, so we need to make sure that we use the correct one for
the argtypes tuple.

I think it would be cleaner to create a `uniffi` library, like we do on rust
and that could hold the `RustBuffer` implementation.  But I'm not sure
it's worth the extra complication for the uniffi user.