Implemented wrapper/external types on Kotlin
In contrast to Python, I made new types for the wrapped types on Kotlin
rather than just using the primitive.  This seems to make more sense for
a typed language, for example maybe the consuming library wants to store
a Guid and wants the type-checker to make ensure that some non-guid
string doesn't get used.

I ran into the same issue with having multiple copies of `RustBuffer`,
`RustBufferBuilder`, etc.  Static typing made it worse than on Python.
I had to define an API for writing to a `RustBufferBuilder` without
actually referencing `RustBufferBuilder`, to work around the issues.

The packaging code works for the tests, but we need to tweak it or
replace it all together to use this in the real world.  While we're
doing that, maybe we could fix the above issue by generating a
`uniffi_core` package that defines `RustBuffer`, `RustBufferBuilder`,
etc. and include in the .jar file for all the other packages to share.

Made FFIConverter an interface, which makes wrapping nicer.  I think
this same wrapping code could be used for a lot more, like String -> an
actual JSON object.