Improve error reporting when an object impl isn't Sync+Send.
We require the implementation of all object structs to be Sync+Send,
and the generated Rust scaffolding will fail to compile if this
requirement is not met. Unfortunately, the failure is currently
triggered via some trait bounds on `FfiConverter` for `Arc<T>`,
so the error message it produces is quite inscrutable and doesn't
help the user debug the underlying issue.

This commit adds an explicit static assertion about the Sync+Send
requirement, which helps the Rust compiler to emit the proper diagnostic
about why the underlying struct is not `Sync` or `Send`. It also has
the added benefit of guarding against any future refactor of the
code that might accidentally factor away the Sync+Send requirement.

Fixes #1048. (Well, not really *fixes* I guess because the code in
experience of debugging such cases much better).