New system for type dispatch
This is my attempt to improve on the `CodeOracle -> CodeType` pattern by using
macros.  The benefits of this are:

  - No need to convert `Type` to something else, the traits get
    implemented on `Type` itself.
  - Only need to define the matching code once.  The language-specific
    `CodeOracle` is replaced by the macros in
    `types::interface::dispatch`.
  - We can split up the traits into a general and languages-specific
    version
  - Also dispatches traits for types that correspond to a `Type` variants (`Record`,
    `Enum`, etc).
  - No `Box<dyn trait>` needed.

Right now only the general logic is implemented.  The next step is to
update the Kotlin code to use it.  I wasn't sure what to name this, and
I didn't want to conflict with the existing name, so I just went with
`NewCodeType`.  If this is successful, then it can replace `CodeType`
and take the name.

While I was updating things, I took the opportunity to tweak
`canonical_name` so that primitive types are uppercase.  I think this
makes for nicer function/class names (OptionalString vs Optionalstring).