Load Python library from absolute path
First of the `*.so` file extension is valid on more than just Linux,
e.g. on BSD and Solaris too.
We can just consider it the default case and only change it for Windows
and macOS.

Next, for final deployment the compiled library can be placed next to
the Python code and that way the wrapper code can find it by its
absolute path.