Implement callback functionality for Python
We use the same pattern as for Swift and Kotlin:
We wrap a type, send over a simple callback function and do the
dispatching to the right instance method on the Python side again.

There's one catch:
`ctypes`' callback definition can't have the callback return a struct by value.
So we change the callback function slightly:

It returns an integer.
When positive it signals success.
It also gets an out pointer to a RustBuffer,
into which it writes the RustBuffer data.

Back on the Rust side a slot for the RustBuffer is set up,
passed in and afterwards used.
That's safe because an empty RustBuffer doesn't allocate, so there's
nothing to deallocate.

Sidenote:
Python's ctypes uses libffi.
There's were known issues with libffi's callbacks.
Current libffi versions should gracefully handle running under
SELinux. Previous versions might just crash when SELinux is enabled.
Another still-existing bug is about the interaction of libffi's callback
(de)allocation, when the application also uses `fork`.
We rely on functions not being GC'ed by Python and thus the callback never being deallocated,
thus not causing that issue for Uniffi-generated code.

Fixes #1068