Implementing wrapped types in the bindings code
The wrapped types are specified in the `uniffi.toml` config file.  This
means the `CodeOracle` structs need to store the config.  Updated the
filters code to handle this.  I wish there was a better way to handle
this, but I couldn't think of one.

I had to disable the external test for Python.  I need to do another
pass on that to make it work well again after the UOC changes.

Re-run uniffi_bindgen if `uniffi.toml` changes.  This is more important
now that `uniffi.toml` stores the wrapper config.

Updated the docs for this.