Switch dispatch system to be more template-centric
Now that Kotlin uses the `FfiConverter` pattern, there's a lot less
logic in the modules in gen_kotlin that correspond to the `Type` variants.
This commit attempts to refactor/simplify things by removing those
modules and moving the functionality to other places.

The main thing those modules were doing was code generation through the
`CodeDeclaration` trait and the `imports()` and `helper_code()` methods.
I replaced that with a new system that's mostly implemented in the
template code.

In the new system `Components.kt` iterates through the types and
includes a template specific to the `Type` variant (kind of like the
`Oracle.find()` method).  It also iterates through the functions and
includes `TopLevelFunctionTemplate.kt` for each of them.  Each of those
templates can do things like add imports/initialization code or include
a template, but only once.

The `type_name()` and `literal()` methods don't fit into that pattern,
so they're back to using plain matches.  `canonical_name()` now uses the
version from the inherent impl.

Implemented `AsRef<Type>` for `Argument` and `Field` so that we can
continue to use them with the template filters.