Allocate RustBuffers all at once (#1152)
Instead of growing the buffer as we write to it, calculate the size
up-front and allocate it all at once.

The main reason for the change is to get feedback on the general
approach because this is how I want to implement things for the upcoming
the desktop JS bindings.  Does this seem like a reasonable approach to
buffer handling?

For the desktop JS bindings, the plan is to use the `ArrayBuffer` class
to pack data into.  However `ArrayBuffer` doesn't support growing the
capacity.  I think it's simpler overall to allocate everything up-front
rather than implementing a growable `ArrayBuffer`.

There are a some side benefits though:
  - It should be more efficient, since you don't to re-allocate or copy
    the data over.
  - The write() methods input `ByteBuffer` rather than
    `RustBufferBuilder`.  Using a standard type is nice when you start
    dealing with external types.  You don't have to worry about the fact
    the external type will have it's own `RustBufferBuilder` type.
  - The `lowerIntoRustBuffer()` code is simpler than before

I also standardized a few things:
  - Made FfiConverter an interface
  - Made all the lift/lower/read/write argument names the same
  - Organized the classes so the functions come in the same order