Doc fixes around custom types
Fixed the CHANGELOG entry for the feature.  It's not released yet, but
it was listed in the changelog as being part of 0.16.0.  Looks like the
wires got crossed because the feature was merged a couple days after
the 0.16.0 release.

Updated the docs to refer to them as "Custom types" instead of "Wrapped
types"